LayoutTests/printing/page-break-always-for-overflow-expected.txt

Test page-break-{before,after}:always for overflow:{visible,hidden,scroll,auto} elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layoutTestController.pageNumberForElementById('test-page-break-before:always-for-overflow:visible') is 1
PASS layoutTestController.pageNumberForElementById('test-page-break-before:always-for-overflow:hidden') is 2
PASS layoutTestController.pageNumberForElementById('test-page-break-before:always-for-overflow:scroll') is 3
PASS layoutTestController.pageNumberForElementById('test-page-break-before:always-for-overflow:auto') is 4
PASS layoutTestController.pageNumberForElementById('test-page-break-after:always-for-overflow:visible') is 4
PASS layoutTestController.pageNumberForElementById('test-page-break-after:always-for-overflow:hidden') is 5
PASS layoutTestController.pageNumberForElementById('test-page-break-after:always-for-overflow:scroll') is 6
PASS layoutTestController.pageNumberForElementById('test-page-break-after:always-for-overflow:auto') is 7
PASS layoutTestController.pageNumberForElementById('test-last-page') is 8
PASS successfullyParsed is true

TEST COMPLETE
